Rewrote normalize_absolute_path
This is a longer implementation, and the root calculation is the longest part due to handling windows (which has a 2-character root for absolute file paths).  Though it's not fully refined, have put this as a pull request so that automated tests can be easily run on the code.

Once code is proven to be working, refinements can be made.