Update package_specs_from_bsconfig and add warnings field.
package_specs_from_bsconfig needed to be updated because for native
we also needed the entries. We're not currently using them though.

Also this adds a `warnings` field that is configurable from the
bsconfig. This doesn't change the default warnings, just makes them
usable for when compiling to native later.

Update docs.

Use `-warn-error` for building deps in release mode.

Updated to use OCAMLPARAM to set -w with warnings field from bsconfig.

This appends the warnings to the default ones, which should allow you
to disable / enable any warning.

I switched to using `Unix.execvpe` for windows too as it seemed like
it was supported. I have _NOT_ tested this yet. Will do when I get home
tonight.

Fixes. Still missing adding the env to the windows command when
building deps.