Update package_specs_from_bsconfig and add warnings field.
package_specs_from_bsconfig needed to be updated because for native
we also needed the entries. We're not currently using them though.

Also this adds a `warnings` field that is configurable from the
bsconfig. This doesn't change the default warnings, just makes them
usable for when compiling to native later.