Super warning & error reporters
This swaps out the default warning & error reporting logic of the compiler for our custom one. The new system is triggered when -better-errors is passed to the compiler.

We forward warning & error reporting to the compiler when we don't handle the specific ones we're concerned with, so this is safe.