Fixed spurious 'configuration failure' message
On non-win32 systems, "configuration failure" is spit out to the console
irregardless of a successful installation.

So, I changed to code to say "configuration failure" iff it fails to find
and use a config_map on win32 (the original intention of the code),
or to say "Non Windows 32 Bit System" if it's not being installed
 on a Windows 32 Bit System
(Much less confusing/ominous than 'configuration failure').

Signed-Off-By: Elliot J. Williams <ejwilliams@wesleyan.edu>