[Experiment] Represent ocaml records as js objects at runtime.
Experiment with representing ocaml records as js objects at runtime, to assess the number of changes involved to support a debug mode whether the field names are visible.

For example `let r = {x=12; y=34}` becomes
```
var a = /* record */{
  x: 12,
  y: 34
};
```

Field access and update are also supported.

This PR changes the core functions plus a few cases that rely on the assumption that arrays are used at runtime.

A few known missing thigs are:
- generic compare always returns 0, as it expects blocks with tag and length
- accessing the content of refs with .contents does not work
- the lexer assumes an array representation
- surely more
But for self-contained code, it’s possible to run some experiments already.