Fix switch optimization incompatible with record access.
The 3 cases in this function are collapsed into one:
```
let getval =
  let f x = x+1 in
  function
  | A {a} -> f a
  | B {b} -> f b
  | C {c} -> f c

```
This is because the index access in the 3 cases is the same: the first index.
When switching to the object representation, this optimization is unsound.

Enriched comparison of lambda primitives to distinguish access to fields with different names, even though they might have the same index.

With this, all tests pass.