Avoid using Objects.keys but use for.in loops.
Found a way to avoid calling `Object.keys entirely` and in principle avoid allocation.
A key is a witness that `a` is smaller than `b` if it's only in `b` or it's in both and the corresponding value in `a` is smaller.
It's possible to traverse `a` and `b` separately to find the smallest (in alphabetical order) key witnessing that `a` is smaller, and that `b` is smaller, then the smaller of the two keys determines the result.