Minimal extension to record representation for debugging.
Minimal extension to the runtime representation of records that could be used for debugging.
A record `{x:12, y:34}` is represented as “[ 12, 34, '.labels': [ 'x', 'y' ] ]”.

This change is minimally invasive compared to a direct representation of records as objects.
At the same time, the representation is more indirect.

Illustrate with a test that with this representation, the conversion between objects and records can be automated.