Try debugging info for variant constructors.
Add debugging info for non-constant variant constructors.
So in
```
type t = A | B of int
```
the debugging will only apply to B. Since A is represented with the constant 0, it would be much more invasive to change representation.

The list `[1;2;3]` is represented as `[ 1, [ 2, [ 3, 0, cstr: '::' ], cstr: '::' ], cstr: '::' ]`. It would be easy to have a custom printer for lists based on this kind of information.