Make changes to belt to enable it for native.
This mostly adds compile time flags to hide all externals that are not
valid in native, and tweaks some of the internally used APIs.

The three big changes are related to arrays, lists and hashing.

For arrays I had to change makeUninitializedUnsafe to take a third
parameter, which is the default value for the array and which is
ignored in JS. That is necessary for native compilation as the
runtime representation of an array of floats is different from an array
of anything else (1 vs 2 words per item). There are multiple ways to
go about this, for example I could've added compile time flags at each
call site, to call a special version of that function for native.
I thought that this would only slow down development on belt because
you'd need to remember that compile time flag each time you use arrays
internally. I did expose a really unsafe version only for JS, which
is used in some places for simplicity (see the bottom of the belt_Array
file).

For Lists I wanted to keep the creation performance and the general
logic of all the functions, which actually assume that the list is
mutable. Having a mutable list isn't something that ocaml generally lets
you do when compiling to native unless you write some C code and you
use the Obj module. The C code is necessary for creating the mutable
list because OCaml tracks mutability for optimization purposes. If we
were to use `::` or `[]` (the list append or list creation literal),
the compiler will assume the resulting list is definitively immutable
which it isn't (at least internally to Belt). What I found out is that
the compiler does not assume immutable when a list is allocated from
an external. So I wrote an external that does exactly what `x :: []`
would do. Then I could "safely" use the Obj module to imperatively /
mutatively set fields of the list. This helps us keep the performance
up for creating lists, which usually would take 2 passes (immutably
append to front and then reverse the list).

For hashing I did a hack job, and I'd love to get some help there.
I copied from a place that I don't remember functions that do what this
does https://github.com/BuckleScript/bucklescript/blob/68f7073d5c41b37703a376837974d52a8e00b886/jscomp/stubs/ext_basic_hash_stubs.c
but in pure ocaml. There might be some funniness about 32bit/64bit
assumptions in there that I can't reason about, knowing nothing about
hashing.