Better fix actually.
We use start for when we're parsing an #if, #elif or #else.
We use end for when we're parsing #end.
We use -1 for when we reach EOF or at the beginning.

Also bspp_main is getting called with linenumbers from after the thing
was parsed, so we use the previous one and set it each loop.