Add `serializeAnyToJson` and `unserializeAnyFromJsonUnsafe`
See docs for more info. This also retains debug information if bsc was
with with `-bs-g`.

```
[anyToJson value] turns any [value] into a JSON object

This will throw an error if there are any functions anywhere in the value.

[unsafeAnyFromJson json] converts a serialized JSON object back into the bucklescript runtime value.

Warning: marshaling is currently not type-safe. The type of marshaled data is not transmitted along
the value of the data, making it impossible to check that the data read back possesses the type expected
by the context. The return type of this function is given as 'a, but this is misleading: the returned
OCaml value does not possess type 'a for all 'a; it has one, unique type which cannot be determined at
compile-time.

The programmer should explicitly give the expected type of the returned value, using the following syntax:

  (Js.Json.unserializeAnyFromJsonUnsafe json : type)

Anything can happen at run-time if the object does not correspond to the assumed type.
```

Test Plan:
```
make libs
make -C jscomp/tests all
./node_modules/.bin/mocha jscomp/test/js_json_test.js
```