Refactor handlers to avoid "double onExit" call
This avoids the fact that both OS signal handlers (SIGINT, SIGUSR, etc)
were calling the same exact function that was also used to handle the "exit" event on
node. Since this function was cleaning the lock file it means that the
second call would just throw an exception because the lock file was
already deleted.

The exit code was also changed to 0 since this is a gracefull shutdown
and no error occurred

Signed-off-by: André Fonseca <mail@andrefonseca.pt>