Add default support for monadic sugar with promises.
This is obviously not ready to be merged yet. But my hope is that, once https://github.com/facebook/reason/pull/2487 is merged, we can get default support for monadic sugar with JS promises in Reason. Expected usage would be like this:

```reason
open Js.Promise;

{
    let.async name = resolve("amazeface");
    Js.log2("The user's name  is " ++ name); 
}
```

Except, I just realized that I don't think this will work, because the source file is in OCaml syntax, and (let.async) isn't valid OCaml syntax. So this PR may not even be possible.

Would you consider the possibility of changing the js_promise bindings source file to be a Reason file after #2748 is merged? I can see that you might be trying to keep Reason out of the source so that pure OCaml users of Bucklescript don't have to mess with Reason. But according to my understanding bsb should be able to compile both Reason and OCaml files in the same project without an issue, right?