Improve printing multiline strings and template literals (#191)
* Improve printing of multiline strings and template literals.

When the content of the string exceeds the line length, the printer breaks the layout. Line breaks in strings itself are just counted as characters now. However they form actual line breaks in the layout. A new type of linebreak `LiteralLine` is introduced to fix it. It is always included and forces breaks in the parents, but doesn't increase the indentation of the subsequent line.

* Don't strip trailing whitespace in multiline string/template content.

This is user space, we can't touch it

* Only split string contents on `\n`.

If we have strings with binary data, \r will become \n after printing. (Changes the meaning of the source)