[Super errors] Remove dependency on Format for the code frame display
This is a mostly feature-compatible refactor of super_misc, which is used to display the code frame in super errors reporting. I've optimized, and added useless features, to keep the tests mostly still intact so that this can be merged more comfortably. After this gets merged, I'll clean up some more logic.

Changes:
- No more Format.
- No more need to patch compiler for colors (will do this in the next diffs).
- Less need to understand the intricacies of coloring. It's now mostly encapsulate within this module.