Fix issue where exhaustiveness check for pattern matching char was not working
Fixes https://github.com/rescript-lang/rescript-compiler/issues/5557
Fixes https://github.com/rescript-lang/rescript-compiler/issues/5743

Exhaustiveness check for pattern matching relies on checking that the pattern does not lead to a type error.
Since char constants are represented internally as int, this causes an internal type error when processing patterns of the form 'x', where the constant is represented as `Pconst_integer` but the expected type is `char`.

This PR intercepts this situation and changes the expected type to `int`.
A cleaner solution would be to represent chars differently (but we can't change the AST), or make it official that char and int is the same type (and lose some abstraction).