Syntax: process uncurried function declarations explicitly in the parser/printer (#5794)
* process uncurried function declarations explicitly in the parser/printer

* Update CHANGELOG.md

* Remove example that is not idempotent

This is not idempotent:
let c2 = (. x) => y => x+y

* Interpred arity of uncurried explicitly.

Interpret the arity of uncurried functions explicitly so this is a unary function:
```
(. x) => y => x+y
```
The parser adds braces to the body to make this explicit.

This is a breaking change. And solves the lack of idempotency in https://github.com/rescript-lang/rescript-compiler/pull/5794

* Update CHANGELOG.md

* Interpred arity of uncurried explicitly.

Interpret the arity of uncurried functions explicitly so this is a unary function:
```
(. x) => y => x+y
```
The parser adds braces to the body to make this explicit.

This is a breaking change. And solves the lack of idempotency in https://github.com/rescript-lang/rescript-compiler/pull/5794

* Update CHANGELOG.md

* snap

* Added analogous tests for types.

* Update CHANGELOG.md