Fix uncurried case of multiple args, and handle corner case.
The corner case has the form:
```res
(x, .y) => z => 3
```

The dot in the middle of a sequence of args does not have a corresponding form in uncurried by default, because non-dots in uncurried by defaults are the normal ways to express the argument after the first.

This means there's asymmetry between the two syntaxes.
This could be handled by declaring this syntax invalid.

Currently we instead parse to the desugaring of the syntax, which in normal mode is:
```res
x => (.y) => { z => 3 }
```

And in uncurried mode is expressed as:
```res
(.x) => y => { (. z) => 3 }
```