Add support for uncurried @this
Both in types in externals and in function definitions, the `@this` annotation now ignores uncurried types. In any case, the processing replaces the type (and the function definition) with some internal representation of method types (and of methods).

This is so when switching to uncurried mode, the code does not need changing.