Fix compiler ppx issue when combining `async` and uncurried application
Fixes https://github.com/rescript-lang/rescript-compiler/issues/5854

The compiler ppx for `async` and `await` was propagating annotations of the body. When the body happens to be an uncurried application, the `@res.app` annotation was propagated to the code generated by the PPX and produced a type error (E.g. it would generate an uncurried call to the internal function `unsafe_await`).