Change representation: pass arity in an attribute.
Using a constraint to determine the arity in a function definition is problematic as the constraint limits type propagation during inference.
Instead, pass the arity via an attribute @res.arity on the Js.Uncurried variant.
And specialize the type checker to handle that arity and use type unification to propagate the arity.