Explore default arguments in uncurried functions without a final unit argument
Functions such as this one

```res
let foo1 = (~x=3, ~y) => x+y
```

 normally require a final unit unlabeled argument to indicate that the optional argument `x` is not passed.

In this example, writing

```res
let r1 = foo1(~y=11)
```

makes `r1` a function that is still expecting the `x` argument.

Requiring a final unit argument, and passing it as `foo1(~y=11, ())` is the only way to omit `x` in  practice.

With uncurried functions, there's the opportunity to treat `foo1(~y=11)` in a special way, meaning that argument `x` is omitted.

TODO:
- Figure out what to do with the warning ""This optional parameter in final position will, in practice, not be optional"
- Figure out the case of all optional arguments. There's no way to pass zero arguments. One could interpret `foo()`, which actually passes `()` as single unlabelled argument, in a special way when `foo` does not accept unlabelled arguments, and use it to mean zero arguments passed.