Fix issue with error messages for uncurried functions where expected and given type were swapped
The logic for unification for `Function$`, which unifies with a generic uncurried type, was swapped.
The error message for arity mismatch was also swapped, so those errors were showing correctly.
However, in the case of expected curried function, when given a curried function whose type is begin inferred, the error message was swapped.
Also added a specialized error message when the uncurried type only contains a type variable, so `$function<'a, ...>` does not leak in the output. This happens as unification failt early, and the type of the function has not been determined yet.