Introduce subtyping instead of ad-hoc treatment of function arguments.
Instead of using an ad-hoc treatment of function arguments which are callbacks in automatic curried application, use subtyping in uncurried-always mode.
So an uncurried function is cast to a curried one during unification.

In addition to the existing examples, this also supports uses of `|>` and callbacks to `div` such as `onClick`, which now can be uncurried functions.