Fix issue with async and locally abstract types
To fix the printer, the async internal representation via an attribute needs to be preserved on functions starting with a locally abstract type.
For this reason, annotations cannot be placed individually on the first locally abstract type, if that's how the function starts.

Fixes https://github.com/rescript-lang/rescript-compiler/issues/5974