Fix issue with JSX V4 when components are nested
Fixes https://github.com/rescript-lang/rescript-compiler/issues/5975

The issue is due to a combination of:
1) when a component let make = body is transformed, first the body is transformed, then when transformStructureItem4 is called, the body is transformed again
2) @react.component is left after the PPX transformation

Because of this, in the case of nested components, the transformed inner components is transformed again by transformStructureItem4, but it should not as it's been desugared already.

It would be enough to change either one, to avoid the error with nested components, but this PR handles both aspects.