Fix implementation of directives.
The compiler's ppx is only run when parsing the source file.
For config settings such as directives, this is OK when `bsc` is invoked directly, but not when part of building a project. In the latter case, the ast is compiled separately, and the builtin ppx is not run again before emitting code.