Try to determine uncurried-always mode in the formatter.
Project config is not used in the formatter normally.
This attempts a best-effort way to find bsconfig.json and see if the mode is set.
The editor currently calls this on a temp file, so it tries to find bsconfig from the path of `bsc.exe` used, by looking outside node_modules.
A more robust method should be provided later on, but this should be OK to start experimenting.