Make intenral encoding of locations aware of unicode
When some unicode characters are present on a line, the existing encoding of positions, based on number of bytes since line start, is incorrect.
This can be seen in e.g. error messages picked up in the editor (or on the command-line).

This PR takes unicode into account. Even thought the ocaml locations are byte-based, one can trick the system by encoding as pos_cnum:
  (number of bytes from file start to line start) + (number of utf16 code units since line start)
Since the compiler's printer performs a subtraction, the utf16 character position is shown.
Notice that editors, vscode in particular, show you something in "Col", but its internal commands expect correct utf16 character which is different.