Pass the information oneUnitArg lower down the compiler stack.
Instead of removing the arguments on the lambda layer, pass the information down via the lambda layer using the additional field oneUnitArg.

When this reaches the Lam layer with ocaml_fun in Lam_compile, only then remove the param.

This ensures the code emitted is the same, except for the parameter.