Infer record vs object at type checking time.
Some syntax is ambiguous:
  `type t = {...t1, ...t2}`
could be either a record or an object, and there's no way to tell.
This PR assumes it is a record and at type-checking time detects if it is not, in which case it type checks the definition as an object.

This is an attempt to get around ambiguity in the syntax that does not seem to be solvable given how crowded the syntax real estate of the language already is.