GenType: fix issue with @as("0") and other numbers.
There is some special code in the type emitter to take care of poly variants of the form e.g. `#0`, which is represented internally as the string "0".
This conflicts with normal variants annotated with `@as("0")` and such.

This PR resolves the case of numeric poly variants early on, and removes the special casing in the code emitter.

Fixes https://github.com/rescript-lang/rescript-compiler/issues/6486