Fix issue with uncurried function with 1 arg being a variable
Fixes https://github.com/rescript-lang/rescript-compiler/issues/6504

Uncurried functions with 1 unit argument are emitted without the parameter. This leaves a possible undefined id in the emitted code if the parameter is a variable.
This PR limits the optimization to cases where the parameter is explicitly `()`, as opposed to e.g. a variable whose type is inferred to be unit.