Workaround for `@as`in labels in uncurried externals.
Examples like this:
```
@obj
external makeOptions: (
  ~objectMode: @as(json`false`) _,
  ~name: string,
  unit,
) => int = ""
```

change the arity of the function.
The arity of uncurried functions is determined at parsing type.
This means that this peculiarity of externals affects parsing.
This PR gives an approximate workaround based on the labelled arg having `@as`, and the type being `_` (the any type).

Fixes https://github.com/rescript-lang/rescript-compiler/issues/6517