Fix issue of incorrect switch cases with identical bodies when mixing object and array.
Fixes https://github.com/rescript-lang/rescript-compiler/issues/6789

The issue happens when 2 cases, here `Object` and `Array`, have identical body (here `Console.log(v)`).
The switch-generation code, which was not designed with untagged unions in mind, merges the two cases into one (and makes one of the two empty).

However, for `Object` and `Array`, what's generated is not a straight switch, but a mix of `if-then-else` and `switch`. This means that the `Object` and `Array` cases are apart in the generated code, and merging them (making one empty) is wrong.

# Conflicts:
#	CHANGELOG.md
#	jscomp/test/UntaggedVariants.js