PPX v4: mark props type in externals as `@live`. (#6796)
* PPX v4: mark props type in externals as `@live`.

Fixes https://github.com/rescript-lang/rescript-vscode/issues/994

Dead code elimination in the editor tooling complains about props never being read in the `props` type record defined by the V4 ppx.
This is because externals don't provide the implementation of the component, and it's in the implementation that props could be read.
This OR marks the `props` type definition as `@live` for external components, so the dead code analysis does not fire.

* Update CHANGELOG.md