Remove coercion with 2 types, which is only supported in ml syntax.
There are two forms of type coercion:
1 `e: t0 :> t`
2 `e :> t`

The first form was never supported in .res syntax, and is now removed from parsed and typed tree.

That said, coercion 1 is the only one that ever supported coercion with free variables. So this is subject to more investigation.

See https://github.com/rescript-lang/rescript-compiler/pull/6828