New mechanism to determine arity of externals.
The arity of external declarations is currently determined just based on the untyped parse tree.
This is probably because the "real" arity is ambiguous in curried mode, and the convention is that an explicit type of the form `t1 => t2 => t3` determines arity 2, but type aliases are ignored.

Here's an example:
https://rescript-lang.org/try?version=v11.1.2&code=C4TwDgpgBArlC8UAUBLAdsANFdwCUCAfDhgFAACAbgIYA2UEAHsBAE5p1SUCMAXLAigAiSkKgB6cVGqsUoKADNWAewC2AlAGccqsLRQBjORRr0mLdp0oAmfqgzZcBeMVyCRYydNnytDRnqGxqS0EMCK3IKMRFzcSIyYjHghYYrWUTE28YnJQA

In the example

```res
type u = (int, int) => int
@val external v1: u = "v" // arity from u is implicit
@val external v2: (int, int) => int = "v" // arity is explicit

let f1 = x => v1(x,x)
let f2 = x => v2(x,x)
```

we have that `v2` is determined to have arity 2, but the arity of `v1` is not determined.
As a consequence, the code generated for `f1` uses runtime `Curry`, both in curried an uncurried mode.

This PR uses the actual type coming from the type checker when in uncurried mode, as that does not suffer from ambiguity.

The consequences are that the external behaves the same however the type is defined, and in the example, no `Curry` runtime is generated.