Fix bug where a ref assignment is moved ouside a conditional
Fixes https://github.com/rescript-lang/rescript/issues/7170

Assignment of a ref, or other record with a single mutable field, is compiled to simple variable assignment.
This triggers an incorrect optimisation that moves identical assignments in the true and false branches of a conditional above if they are identical and the guard has no side effects.

Added a check that if the assigments to be moved are to variable, that variable must not occur free in the guard of the conditional.